add drawCues
