+++ src/listeners/shortcuts/beat-select.js
@@ -4,7 +4,7 @@ var songManager = require('../../config/song-manager.js')
 var beatArray = require('../../config/beat-array.js')
 var showMoveInfo = require('../../moves/show-move-info.js')
 import { createNote } from '../../stage/elements/create-note.js'
-import drawStaticCues from '../../stage/cueContainer/draw-static-cues.js'
+import drawCues from '../../stage/cueContainer/draw-cues.js'
 
 // seek takes some time to take effect for some reason. setTimeout is here to give the desired effect of note creation
 
@@ -16,7 +16,7 @@ function beatSelect (event) {
       createNoteWhenSelected(skippedBeats)
       setTimeout(function () {
         showMoveInfo()
-        drawStaticCues()
+        drawCues()
       }, 150)
     } else if ( event.key === editor.shortCuts.previousQuarterBeat ) {
       let skippedBeats = -1
@@ -24,7 +24,7 @@ function beatSelect (event) {
       createNoteWhenSelected(skippedBeats)
       setTimeout(function () {
         showMoveInfo()
-        drawStaticCues()
+        drawCues()
       }, 150)
     } else if ( event.key === editor.shortCuts.nextBeat ) {
       let skippedBeats = 4
@@ -32,7 +32,7 @@ function beatSelect (event) {
       createNoteWhenSelected(skippedBeats)
       setTimeout(function () {
         showMoveInfo()
-        drawStaticCues()
+        drawCues()
       }, 150)
     } else if ( event.key === editor.shortCuts.previousBeat ) {
       let skippedBeats = -4
@@ -40,7 +40,7 @@ function beatSelect (event) {
       createNoteWhenSelected(skippedBeats)
       setTimeout(function () {
         showMoveInfo()
-        drawStaticCues()
+        drawCues()
       }, 150)
     }
   }
