+++ src/stage/cueContainer/add-cue-graphics.js
@@ -1,0 +1,10 @@
+import { cueContainer } from '../../config/containers.js'
+import * as PIXI from 'pixi.js'
+
+function addCueGraphics() {
+  var cues = new PIXI.Graphics()
+  cues.name = 'cues'
+  cueContainer.addChild(cues)
+}
+
+export { addCueGraphics }
